[UPDATE] v1.5 : Fix average cart request & optimize sub header request

+++ LonaciAPI-web/src/main/java/sn/intouch/gu/lonaciapi/ws/dto/SubHeaderResponse.java
@@ -6,14 +6,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.util.Map;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class SubHeaderResponse implements Serializable {
-    private Integer operationsNumber;
+    private Long operationsNumber;
     private Integer activeClients;
     private Double averageCart;
     private Double overallVolume;
